// LAP/backend/src/common/enums/workflow-log-action.enum.ts
export enum WorkflowLogAction {
  LEAD_CREATED = 'LEAD_CREATED',
  LEAD_SUBMITTED = 'LEAD_SUBMITTED',
  CUSTOMER_VISIT_DONE = 'CUSTOMER_VISIT_DONE',
  BUSINESS_VISIT_DONE = 'BUSINESS_VISIT_DONE',
  GEO_VERIFICATION_DONE = 'GEO_VERIFICATION_DONE',
  PROPERTY_VISIT_DONE = 'PROPERTY_VISIT_DONE',
  DOCUMENTS_UPLOADED = 'DOCUMENTS_UPLOADED',
  SUBMITTED_TO_BM = 'SUBMITTED_TO_BM',
}